Move PlayerItem one package level higher
